+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [3.0.9] - 2025-07-14
+
+### Added
+
+- Hook needed for Global Section setting and for a Section specific theme settings #207
+
+### Fixed
+
+- Guide not highlighting the options correctly: When you are using a layout the guide is not highlighting the options correctly #202
+
 ### [3.0.8] - 2025-07-09
 
 ### Fixed